refactor(cli): extract asset subcommands into commands/ module
Start splitting the 5949-line main.rs into per-command modules. First:
the `asset` subcommands (create/init/verify/add-peripheral, ~250 lines)
move to crates/cli/src/commands/asset.rs. Establishes the cli/commands/
module; EXIT_* consts + resolve_chip_descriptor_path made pub(crate).

main.rs 5949 -> 5699. Pure move — labwired-cli build + clippy clean.
