+++ crates/cli/src/commands/mod.rs
@@ -1,0 +1,3 @@
+//! CLI subcommand handlers, split out of `main.rs`.
+
+pub mod asset;
